serial: use this app as a secure radio transmitter
This app retrieves SECURE data, so its only mission (for now) is to transmit those data. Currently there is only one legitimate receiver, the health app, so a simple HTTP REST call will be transmitted instead of broadcasting or sending the data to a comunications broker.

Note also that the data is not stored here.