docker: first version of docker index with node readline (working but not as expected)
Just for backup purposes, because node readline is intended to emulate GNU readline and that is like an util for CLI applications. And when the stream is a file, the readline does not work as anybody should expect. It can swallow several lines before invoking the 'line' event, so it process line by line yes but you can not process it 100% sequentially